Update Robot.cpp
I didn't have the latest and I can't test the string addition, but it should be close.  Also check the file name "/home/lvuser/TestFile.csv" as I wasn't sure if that is in the right directory.

Anyway this should put all the data in a single file and then you can graph the points in excel and see based on the graphs which items are not consistent.

Also doing the string may help with the chopped file sizes as when we looked at it it would only go to 7.xxxx seconds.  So either the buffer needed to be flushed or it just maxed out.  So I hope with this it allows for the string to grow and then dump it to a single file.